refactor: Add server API types
Add types 'User' and 'ResponseBody', The 'User' type will be used on both-sides
the client and server, 'ResponseBody' used for 'api/v2/users' so every
types for response methods will be added here in this file.

Also 'RequestMethod' and 'RequestMethodsManager' both types is a main types that
should be used to handle incoming requests or when creating new endpoints.