Add new API routes for managing products.
This commit will also include the implementation of 'getAllProductsFromDB' and 'getAllCategoriesFromDB' functions, which can be accessed through the same API routes '/v2/products' and '/v2/categories' but with a 'only_popular' query parameter.
If this parameter is not included and not set to 'true', the API will return all products or categories.